+++ _config.yml
@@ -1,3 +1,4 @@
 exclude:
   - 'node_modules'
   - 'readme.md'
+  - 'page.js'
