+++ _layouts/index.html
@@ -23,8 +23,6 @@
     });
   </script>
   <script src="https://code.jquery.com/jquery-2.1.3.min.js"></script>
-  <script src="prettyprint.js"></script>
-  <script src="plastiq.js"></script>
   <script src="app.js"></script>
   <a href="https://github.com/featurist/plastiq"><img style="position: absolute; top: 0; right: 0; border: 0;" src="https://camo.githubusercontent.com/52760788cde945287fbb584134c4cbc2bc36f904/68747470733a2f2f73332e616d617a6f6e6177732e636f6d2f6769746875622f726962626f6e732f666f726b6d655f72696768745f77686974655f6666666666662e706e67" alt="Fork me on GitHub" data-canonical-src="https://s3.amazonaws.com/github/ribbons/forkme_right_white_ffffff.png"></a>
 </body>
